# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Librarie API",
        "version": "0.1"
    },
    "paths": {
...
        "/cos": {
            "put": {
                "tags": [
                    "CarteCos"
                ],
                "summary": "Update cantitatea cartii din cos",
                "description": "Update cantitatea cartii din cos",
                "operationId": "9ac2df0fd4d7f21b6bdc6273bd56bf47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cantitate": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "CarteCos"
                ],
                "summary": "Adauga o carte in cos",
                "description": "Adauga o carte in cos",
                "operationId": "78a08161d32a482bbbe098eb1047d691",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_carte": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantitate": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subtotal": {
                                        "description": "",
                                        "type": "float",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/cos/{id}": {
            "delete": {
                "tags": [
                    "CarteCos"
                ],
                "summary": "Sterge carte cu id-ul dat din cos",
                "description": "Sterge carte cu id-ul dat din cos",
                "operationId": "6f17c086657c5be8b3e6164d133c146c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul cartii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/categorii": {
            "get": {
                "tags": [
                    "Categorii"
                ],
                "summary": "Returneaza toate categoriile",
                "description": "Returneaza toate categoriile",
                "operationId": "19335a1add68897ebbc80ec81c51677f",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorii"
                ],
                "summary": "Update categoria cu id-ul dat",
                "description": "Update categorii cu id-ul dat",
                "operationId": "05b5cc2560fcc84885911e63faa017b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nume": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorii"
                ],
                "summary": "Adauga categoria",
                "description": "Adauga categorii",
                "operationId": "d7d9db6d591138bb42aee88ee8af6dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nume": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/categorii/{id}": {
            "get": {
                "tags": [
                    "Categorii"
                ],
                "summary": "Returneaza categoria cu id-ul dat",
                "description": "Returneaza categoria cu id-ul dat",
                "operationId": "0784657f7848dd1fc0349ba9dc5282fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul categoriei dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorii"
                ],
                "summary": "Sterge categoria cu id-ul dat",
                "description": "Sterge categoria cu id-ul dat",
                "operationId": "56b5d019b573b1c56a7a4b1fee3e7ff7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul categoriei dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/categorie/carti/{id}": {
            "get": {
                "tags": [
                    "Categorii"
                ],
                "summary": "Returneaza cartile din categoria cu id-ul dat",
                "description": "Returneaza cartile din categoria cu id-ul dat",
                "operationId": "a6404f099827f7570183e2a81a4be205",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul categoriei dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}